+++ CHANGELOG.md
@@ -16,6 +16,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Fixed
+
+- gracefully deal with permission exception (#32)
+
 ## [1.3.0] - 2023-04-03
 
 ### Fixed